Fix unit tests

+++ src/SFA.DAS.EmployerApprenticeshipsService.Application/Mappings/LegalEntityMappings.cs
@@ -56,7 +56,7 @@ namespace SFA.DAS.EAS.Application.Mappings
                     l.LegalEntity.Source == OrganisationType.CompaniesHouse ? "Companies House" :
                     l.LegalEntity.Source == OrganisationType.Charities ? "Charities" :
                     l.LegalEntity.Source == OrganisationType.PublicBodies ? "Public Bodies" : "Other"))
-                .ForMember(d => d.SourceNumeric, o => o.MapFrom(l => l.LegalEntity.Source))
+                .ForMember(d => d.SourceNumeric, o => o.MapFrom(l => (short) l.LegalEntity.Source))
                 .ForMember(d => d.AgreementSignedByName, o => o.Ignore())
                 .ForMember(d => d.AgreementSignedDate, o => o.Ignore())
                 .ForMember(d => d.AgreementStatus, o => o.Ignore());
